+++ src/powergim/testcases.py
@@ -3,6 +3,8 @@ import pandas as pd
 
 from powergim.grid_data import GridData as grid
 
+from . import utils
+
 
 def create_case(investment_years, number_nodes, number_timesteps, base_MW=200):
     """Create test case of a given size"""
@@ -51,7 +53,6 @@ def create_case(investment_years, number_nodes, number_timesteps, base_MW=200):
         else:
             consumers[f"demand_{year}"] = 0
     consumers["demand_ref"] = "demand1"
-    consumers["emission_cap"] = -1
 
     generators = pd.DataFrame(columns=cols_generator)
     # generators at odd number nodes
@@ -134,10 +135,12 @@ def create_case(investment_years, number_nodes, number_timesteps, base_MW=200):
             "finance_years": 40,
             "operation_maintenance_rate": 0.05,
             "CO2_price": 0,
+            "CO2_cap": None,
             "load_shed_penalty": 10000,
             "profiles_period_suffix": False,
         },
     }
+    utils.validate_parameter_data(parameter_data)
 
     return grid_data, parameter_data
 
@@ -196,7 +199,6 @@ def create_case_star(investment_years, number_nodes, number_timesteps, base_MW=2
         else:
             consumers[f"demand_{year}"] = 0
     consumers["demand_ref"] = "demand1"
-    consumers["emission_cap"] = -1
 
     # generators on every second node on top half, at least 2
     if number_nodes >= 7:
@@ -286,9 +288,11 @@ def create_case_star(investment_years, number_nodes, number_timesteps, base_MW=2
             "finance_years": 40,
             "operation_maintenance_rate": 0.05,
             "CO2_price": 0,
+            "CO2_cap": None,
             "load_shed_penalty": 10000,
             "profiles_period_suffix": False,
         },
     }
+    utils.validate_parameter_data(parameter_data)
 
     return grid_data, parameter_data